Remove SetDeadline from the framing layer (#21)
The framing layer is not supposed to have any blocking operations or operations which may hang for a long time. So having methods/parameters related to any kind of timeouts or deadlines looks insane and the framing layer should not manage them in any way. It should just reformat written bytes and join bytes before reading.