Refactor the call protocol (#23)
Remaining response bytes returned by the call protocol is used now only to generate an error. But the call protocol is able to do it by itself so its interface may be simplified what was done in this PR.

Additionally, all call protocol tests were refactored to use regular mocks and to be more accurate.