Extract transports into a separated package (#25)
Now the whole code of the project is placed in a single package. But it contains the code to solve different tasks like transferring bytes in different ways (over sockets, buffered transfer, transfer compressed bytes and so on), implementing different protocols between services (avro message protocol, call protocol) and finally providing a completed client to join together all those stuff. As a result, it may be hard to explore the code and to understand where any part of the whole client is placed. So having different packages for the code to solve different kind of task looks sane in this case.