Flush data from buffer on closing (#33)
Now there is a bug with the buffered transport and if there are some bytes in the buffer before closing the transport they will be silently ignored. It does not create any problem with communication with a Flume server because there is no closing procedure and we can just loose only some unnecessary data of zlib transport. That's all.